@@ -6,7 +6,7 @@
     "build": "next build",
     "build:analyze": "DEV_ANALYZE_BUNDLE=true next build",
     "build:analyze:bundlesize": "cd ../.. && pnpm build --filter=@bifrost-starter/frontend && cd - && bundlesize",
-    "dev": "./gen_env.sh local && next dev -p 3000",
+    "dev": "./gen_env.sh local && next dev --turbo -p 3000",
     "lint": "eslint --cache --cache-location .next/cache/eslint/ './**/*.{js,ts,tsx}'",
     "lint:fix": "pnpm lint --fix",
     "lint:style": "stylelint './**/*.module.css'",
@@ -20,20 +20,22 @@
   },
   "dependencies": {
     "@bifrost-starter/interfaces": "workspace:*",
-    "@next/font": "13.4.7",
     "axios": "0.27.2",
     "axios-auth-refresh": "3.3.6",
+    "client-only": "^0.0.1",
     "jwt-decode": "3.1.2",
-    "next": "13.4.7",
+    "next": "14.0.4",
+    "next-intl": "^3.3.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "7.45.1",
-    "react-intl": "6.4.4",
-    "swr": "1.3.0"
+    "server-only": "^0.0.1",
+    "sharp": "0.32.6",
+    "swr": "2.2.4"
   },
   "devDependencies": {
     "@axe-core/react": "4.7.3",
-    "@next/bundle-analyzer": "13.4.7",
+    "@next/bundle-analyzer": "14.0.4",
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "13.4.0",
     "@types/jest": "29.5.2",
