@@ -1,9 +1,10 @@
+import 'client-only';
 import axios from 'axios';
 import createAuthRefreshInterceptor, {
   AxiosAuthRefreshRequestConfig,
 } from 'axios-auth-refresh';
 import jwtDecode from 'jwt-decode';
-import Router from 'next/router';
+import type { useRouter } from 'next/navigation';
 
 import { Pages } from 'constant';
 import { env } from 'services/env';
@@ -16,59 +17,63 @@ import {
   setAccessToken,
 } from './auth/utils';
 
-export const apiClient = axios.create({
-  baseURL: env('NEXT_PUBLIC_API_BASE_URL'),
-  withCredentials: true,
-  headers: {
-    'Content-Type': 'application/json',
-  },
-});
-
-apiClient.interceptors.request.use(
-  async config => {
-    const access = getAccessToken();
-    if (access === null) {
-      await Router.push(Pages.Login);
-
-      return config;
-    }
-    if (isTokenExpired(jwtDecode(access))) {
-      await refreshToken();
-    }
-
-    return {
-      ...config,
-      headers: Object.assign(config.headers ?? {}, {
-        Authorization: `Bearer ${getAccessToken() as string}`,
-      }),
-    };
-  },
-  undefined,
-  {
-    runWhen: config =>
-      config.url !== ApiRoutes.refresh && config.url !== ApiRoutes.login,
-  },
-);
-
-const refreshToken = async (): Promise<void> => {
-  const options: AxiosAuthRefreshRequestConfig = {
-    skipAuthRefresh: true,
+export const apiClientFactory = (router: ReturnType<typeof useRouter>) => {
+  const apiClient = axios.create({
+    baseURL: env('NEXT_PUBLIC_API_BASE_URL'),
     withCredentials: true,
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  });
+  apiClient.interceptors.request.use(
+    async config => {
+      const access = getAccessToken();
+      if (access === null) {
+        router.push(Pages.Login);
+
+        return config;
+      }
+      if (isTokenExpired(jwtDecode(access))) {
+        await refreshToken();
+      }
+
+      return {
+        ...config,
+        headers: Object.assign(config.headers ?? {}, {
+          Authorization: `Bearer ${getAccessToken() as string}`,
+        }),
+      };
+    },
+    undefined,
+    {
+      runWhen: config =>
+        config.url !== ApiRoutes.refresh && config.url !== ApiRoutes.login,
+    },
+  );
+  const refreshToken = async (): Promise<void> => {
+    const options: AxiosAuthRefreshRequestConfig = {
+      skipAuthRefresh: true,
+      withCredentials: true,
+    };
+
+    try {
+      setAccessToken(
+        getAccessFromResponse(
+          await apiClient.post<unknown>(ApiRoutes.refresh, undefined, options),
+        ),
+      );
+    } catch (error) {
+      router.push(Pages.Login);
+
+      return Promise.reject(error);
+    }
   };
 
-  try {
-    setAccessToken(
-      getAccessFromResponse(
-        await apiClient.post<unknown>(ApiRoutes.refresh, undefined, options),
-      ),
-    );
-  } catch (error) {
-    await Router.push(Pages.Login);
+  createAuthRefreshInterceptor(apiClient, async () => await refreshToken(), {
+    statusCodes: [401],
+  });
 
-    return Promise.reject(error);
-  }
+  return apiClient;
 };
 
-createAuthRefreshInterceptor(apiClient, async () => await refreshToken(), {
-  statusCodes: [401],
-});
+export type apiClientType = ReturnType<typeof apiClientFactory>;
