@@ -1,13 +1,16 @@
 import { getAccessFromResponse, setAccessToken } from './utils';
 import { ApiRoutes } from '../apiRoutes';
-import { apiClient } from '../client';
+import { apiClientType } from '../client';
 
 export type LoginData = {
   email: string;
   password: string;
 };
 
-export const login = async (data: LoginData): Promise<void> =>
+export const login = async (
+  apiClient: apiClientType,
+  data: LoginData,
+): Promise<void> =>
   setAccessToken(
     getAccessFromResponse(
       await apiClient.post<unknown>(ApiRoutes.login, data, {
